Interfaz: Productos terminada con algunos ajutes en Compras y Proveedores

+++ src/main/java/com/eigsacompras/controlador/ProductoControlador.java
@@ -1,23 +1,51 @@
 package com.eigsacompras.controlador;
 
 import com.eigsacompras.dao.ProductoDAO;
+import com.eigsacompras.dao.ProductoProveedorDAO;
 import com.eigsacompras.modelo.Producto;
+import com.eigsacompras.modelo.ProductoProveedor;
+
 import javax.swing.*;
 import java.util.List;
 
 public class ProductoControlador {
     private ProductoDAO productoDAO;
+    private ProductoProveedorDAO productoProveedorDAO;
 
     public ProductoControlador(){
         this.productoDAO = new ProductoDAO();
+        this.productoProveedorDAO = new ProductoProveedorDAO();
     }
 
-    public void agregarProducto(String descripcion,double precio){
-        if(validarProducto(descripcion,precio)){
-            Producto producto = new Producto(precio,descripcion);
-            productoDAO.agregarProducto(producto);
+    public boolean agregarProducto(String descripcion, List<ProductoProveedor> productosProveedores){
+        if(validarProducto(descripcion)){
+            Producto producto = new Producto(descripcion);
+            if(!productosProveedores.isEmpty()){
+                try {
+                    int idProducto = productoDAO.agregarProducto(producto);
+                    if(idProducto!=-1){
+                        for(ProductoProveedor productoProveedor: productosProveedores){
+                            productoProveedor.setIdProducto(idProducto);
+                            productoProveedorDAO.agregarProductoProveedor(productoProveedor);
+                        }
+                        JOptionPane.showMessageDialog(null, "Producto agregado correctamente.", "Agregado", JOptionPane.INFORMATION_MESSAGE);
+                        return true;
+                    }else{
+                        JOptionPane.showMessageDialog(null, "Producto no agregado.\n Si el problema persiste, contacte al desarrollador ", "No agregado", JOptionPane.ERROR_MESSAGE);
+                        return false;
+                    }//if -1
+                } catch (Exception e) {
+                    JOptionPane.showMessageDialog(null, "Error al agregar el producto. Por favor, inténtelo nuevamente. Error: " + e, "No agregado", JOptionPane.ERROR_MESSAGE);
+                    return false;
+                }//try
+            }else {
+                JOptionPane.showMessageDialog(null, "Agregue al menos 1 proveedor.", "Sin productos", JOptionPane.WARNING_MESSAGE);
+                return false;
+            }//if productos proveedor vacio
+
         }else {
-            JOptionPane.showMessageDialog(null,"Hay uno o más campos vacíos, Revíselos","Campo vacío",JOptionPane.WARNING_MESSAGE);
+            JOptionPane.showMessageDialog(null,"Se necesita la descripción del producto","Campo vacío",JOptionPane.WARNING_MESSAGE);
+            return false;
         }
     }//agregar
 
@@ -25,22 +53,55 @@ public class ProductoControlador {
         return productoDAO.listarProducto();
     }//listar
 
-    public void actualizarProducto(String descripcion, double precio, int idProducto){
-        if(validarProducto(descripcion,precio)){
-            Producto producto = new Producto(idProducto,precio,descripcion);
-            productoDAO.agregarProducto(producto);
+    public boolean actualizarProducto(String descripcion, int idProducto, List<ProductoProveedor> productoProveedores){
+        if(validarProducto(descripcion)){
+            Producto producto = new Producto(idProducto,descripcion);
+            if(!productoProveedores.isEmpty()){
+                if(productoDAO.actualizarProducto(producto)){
+                    productoProveedorDAO.eliminarProductoProveedorPorIdProducto(idProducto);//se eliminan los productos con el proveedor
+                    for(ProductoProveedor productoProveedor: productoProveedores){//se agregan los productos nuevamente incluidos los nuevos a ese proveedor
+                        productoProveedorDAO.agregarProductoProveedor(productoProveedor);
+                    }
+                    //se eliminan todos los productos asociados al proveedor ya que si se modifica el proveedor y se quieren agregar otros productos si solo se actualiza no se podrá agregar lo nuevos productos al proveedor...
+                    //...por lo que para asegurar que se actualice correctamente incluyendo los nuevos productos que se quieren agregar es necesario eliminar todo y agregar todo lo nuevo
+                    JOptionPane.showMessageDialog(null, "Todo actualizado correctamente.", "Actualizado", JOptionPane.INFORMATION_MESSAGE);
+                    return true;
+                }else {
+                    JOptionPane.showMessageDialog(null, "Error al actualizar el producto. Por favor, inténtelo nuevamente.", "No actualizado", JOptionPane.ERROR_MESSAGE);
+                    return false;
+                }//if dao
+            }else{
+                JOptionPane.showMessageDialog(null, "Agregue al menos 1 producto.", "Sin productos", JOptionPane.WARNING_MESSAGE);
+                return false;
+            }//if valida productoProveedor vacio
         }else {
-            JOptionPane.showMessageDialog(null,"Hay uno o más campos vacíos, Revíselos","Campo vacío",JOptionPane.WARNING_MESSAGE);
+            JOptionPane.showMessageDialog(null,"Se necesita la descripción del producto","Campo vacío",JOptionPane.WARNING_MESSAGE);
+            return false;
         }
     }//actualizar
 
     public void eliminarProducto(int idProducto){
-        productoDAO.eliminarProducto(idProducto);
+        int opc = JOptionPane.showConfirmDialog(null, "¿Está seguro de eliminar este producto?", "Confirmacion", JOptionPane.YES_NO_OPTION);
+        if (opc == JOptionPane.YES_OPTION) {
+            if(productoProveedorDAO.eliminarProductoProveedorPorIdProducto(idProducto)) {
+                productoDAO.eliminarProducto(idProducto);
+                JOptionPane.showMessageDialog(null, "Producto eliminado correctamente.", "Eliminado", JOptionPane.INFORMATION_MESSAGE);
+            } else {
+                JOptionPane.showMessageDialog(null, "Error al eliminar el producto. Por favor, inténtelo nuevamente.", "No eliminado", JOptionPane.ERROR_MESSAGE);
+            }
+        }
     }//eliminar
 
-    public boolean validarProducto(String descripcion, double precio){
+    public Producto listarProductoPorId(int idProducto){
+        return productoDAO.buscarProductoPorId(idProducto);
+    }//listar por id
+
+    public List<Producto> buscarProductos(String termino){
+        return productoDAO.buscarProductos(termino);
+    }//buscar productos
+
+    public boolean validarProducto(String descripcion){
         if(descripcion.isEmpty()) return false;
-        if(precio == 0) return false;
 
         return true;
     }//validar
